Refactor pipeline stage runtime management
