+++ src/bioetl.egg-info/SOURCES.txt
@@ -18,11 +18,9 @@ src/bioetl/config/pipeline_config_schema.py
 src/bioetl/application/pipelines/__init__.py
 src/bioetl/application/pipelines/base.py
 src/bioetl/application/pipelines/contracts.py
-src/bioetl/application/pipelines/error_policy_facade.py
 src/bioetl/application/pipelines/hooks_impl.py
-src/bioetl/application/pipelines/hooks_registry.py
+src/bioetl/application/pipelines/stage_runtime_manager.py
 src/bioetl/application/pipelines/registry.py
-src/bioetl/application/pipelines/stage_runner_facade.py
 src/bioetl/application/pipelines/chembl/__init__.py
 src/bioetl/application/pipelines/chembl/base.py
 src/bioetl/application/pipelines/chembl/extractor.py
