+++ tests/bioetl/application/pipelines/test_pipeline_collaborators.py
@@ -5,12 +5,8 @@ from unittest.mock import MagicMock
 import pandas as pd
 import pytest
 
-from bioetl.application.pipelines.error_policy_facade import ErrorPolicyFacade
 from bioetl.application.pipelines.hooks_impl import ContinueOnErrorPolicyImpl
-from bioetl.application.pipelines.hooks_registry import HooksRegistry
-from bioetl.application.pipelines.stage_runner_facade import (
-    StageRunnerFacade as StageRunner,
-)
+from bioetl.application.pipelines.stage_runtime_manager import StageRuntimeManager
 from bioetl.domain.errors import PipelineStageError
 from bioetl.domain.models import RunContext, StageResult
 from bioetl.domain.pipelines.contracts import PipelineHookABC
@@ -22,13 +18,15 @@ def _build_context() -> RunContext:
 
 
 @pytest.mark.unit
-def test_hooks_registry_notifies_hooks(mock_logger):
+def test_stage_runtime_notifies_hooks(mock_logger):
     hook = MagicMock(spec=PipelineHookABC)
-    manager = HooksRegistry(
+    manager = StageRuntimeManager(
         logger=mock_logger,
         provider_id=ProviderId("chembl"),
         entity_name="entity",
         hooks=[hook],
+        error_policy=ContinueOnErrorPolicyImpl(),
+        default_on_skip=lambda _: None,
     )
     context = _build_context()
     stage_result = StageResult(
@@ -49,55 +47,37 @@ def test_hooks_registry_notifies_hooks(mock_logger):
 
 
 @pytest.mark.unit
-def test_error_policy_facade_retry_and_skip(mock_logger):
-    hooks_manager = HooksRegistry(
-        logger=mock_logger,
-        provider_id=ProviderId("chembl"),
-        entity_name="entity",
-    )
+def test_stage_runtime_retry_and_skip(mock_logger):
     policy = ContinueOnErrorPolicyImpl(max_retries=1)
-    manager = ErrorPolicyFacade(
-        error_policy=policy,
-        hooks_manager=hooks_manager,
+    manager = StageRuntimeManager(
         logger=mock_logger,
         provider_id=ProviderId("chembl"),
         entity_name="entity",
+        error_policy=policy,
         default_on_skip=lambda stage: f"skipped-{stage}",
     )
     context = _build_context()
 
     action = MagicMock(side_effect=[ValueError("temporary"), "ok"])
-    assert manager.execute("extract", context, action) == "ok"
+    assert manager.execute_stage("extract", context, action) == "ok"
     assert action.call_count == 2
     assert manager.last_error is None
 
     failing_action = MagicMock(side_effect=RuntimeError("boom"))
-    result = manager.execute("transform", context, failing_action)
+    result = manager.execute_stage("transform", context, failing_action)
     assert result == "skipped-transform"
     assert isinstance(manager.last_error, PipelineStageError)
     assert "boom" in manager.get_last_error_messages()[-1]
 
 
 @pytest.mark.unit
-def test_stage_runner_process_and_failure(mock_logger):
-    hooks_manager = HooksRegistry(
+def test_stage_runtime_process_and_failure(mock_logger):
+    manager = StageRuntimeManager(
         logger=mock_logger,
         provider_id=ProviderId("chembl"),
         entity_name="entity",
-    )
-    manager = ErrorPolicyFacade(
         error_policy=ContinueOnErrorPolicyImpl(),
-        hooks_manager=hooks_manager,
-        logger=mock_logger,
-        provider_id=ProviderId("chembl"),
-        entity_name="entity",
         default_on_skip=lambda stage: f"skipped-{stage}",
-    )
-    runner = StageRunner(
-        hooks_manager=hooks_manager,
-        error_policy_facade=manager,
-        entity_name="entity",
-        provider_id=ProviderId("chembl"),
     )
     context = _build_context()
     raw_chunk = pd.DataFrame({"id": [1]})
@@ -110,7 +90,7 @@ def test_stage_runner_process_and_failure(mock_logger):
         validate_started,
         validate_chunks,
         validate_count,
-    ) = runner.process_chunk(
+    ) = manager.process_chunk(
         raw_chunk,
         context,
         transform_started=False,
@@ -131,7 +111,7 @@ def test_stage_runner_process_and_failure(mock_logger):
     assert transform_count == 1 and validate_count == 1
     assert len(validated_chunks) == 1
 
-    transform_stage = runner.make_stage_result(
+    transform_stage = manager.make_stage_result(
         "transform", transform_count, chunks=transform_chunks
     )
     assert transform_stage.records_processed == 1
@@ -145,31 +125,19 @@ def test_stage_runner_process_and_failure(mock_logger):
         run_id=context.run_id,
         cause=RuntimeError("fail"),
     )
-    failure = runner.handle_stage_failure("validate", [transform_stage], context)
+    failure = manager.handle_stage_failure("validate", [transform_stage], context)
     assert not failure.success
     assert failure.errors
 
 
 @pytest.mark.unit
-def test_stage_runner_handles_skip_and_counts(mock_logger):
-    hooks_manager = HooksRegistry(
+def test_stage_runtime_handles_skip_and_counts(mock_logger):
+    manager = StageRuntimeManager(
         logger=mock_logger,
         provider_id=ProviderId("chembl"),
         entity_name="entity",
-    )
-    manager = ErrorPolicyFacade(
         error_policy=ContinueOnErrorPolicyImpl(),
-        hooks_manager=hooks_manager,
-        logger=mock_logger,
-        provider_id=ProviderId("chembl"),
-        entity_name="entity",
         default_on_skip=lambda stage: pd.DataFrame(),
-    )
-    runner = StageRunner(
-        hooks_manager=hooks_manager,
-        error_policy_facade=manager,
-        entity_name="entity",
-        provider_id=ProviderId("chembl"),
     )
     context = _build_context()
     validated_chunks: list[pd.DataFrame] = []
@@ -181,7 +149,7 @@ def test_stage_runner_handles_skip_and_counts(mock_logger):
         validate_started,
         validate_chunks,
         validate_count,
-    ) = runner.process_chunk(
+    ) = manager.process_chunk(
         pd.DataFrame({"id": [1]}),
         context,
         transform_started=False,
@@ -204,32 +172,20 @@ def test_stage_runner_handles_skip_and_counts(mock_logger):
 
 
 @pytest.mark.unit
-def test_stage_runner_raises_on_missing_result(mock_logger):
-    hooks_manager = HooksRegistry(
+def test_stage_runtime_raises_on_missing_result(mock_logger):
+    manager = StageRuntimeManager(
         logger=mock_logger,
         provider_id=ProviderId("chembl"),
         entity_name="entity",
-    )
-    manager = ErrorPolicyFacade(
         error_policy=ContinueOnErrorPolicyImpl(),
-        hooks_manager=hooks_manager,
-        logger=mock_logger,
-        provider_id=ProviderId("chembl"),
-        entity_name="entity",
         default_on_skip=lambda stage: pd.DataFrame(),
-    )
-    runner = StageRunner(
-        hooks_manager=hooks_manager,
-        error_policy_facade=manager,
-        entity_name="entity",
-        provider_id=ProviderId("chembl"),
     )
     context = _build_context()
 
-    manager.execute = MagicMock(return_value=None)  # type: ignore[method-assign]
+    manager.execute_stage = MagicMock(return_value=None)  # type: ignore[method-assign]
 
     with pytest.raises(PipelineStageError):
-        runner.process_chunk(
+        manager.process_chunk(
             pd.DataFrame({"id": [1]}),
             context,
             transform_started=False,
@@ -245,4 +201,4 @@ def test_stage_runner_raises_on_missing_result(mock_logger):
             validate_fn=lambda df: df,
         )
 
-    manager.execute.assert_called_once()
+    manager.execute_stage.assert_called_once()
